update runall.sh
* add rendering wtih blender engine
* specify output name with -o

+++ runall.sh
@@ -1,42 +1,54 @@
+ROOT=$(dirname "$(realpath "$BASH_SOURCE")")
 RESULT=$(pwd)/result
-ROOT=$(dirname "$BASH_SOURCE")
 
 
 function breakline(){
+	echo ""
 	echo "---------- ---------- ----------"
 }
 
+# for state
 function bold_magenta(){
 	echo -e "\033[35m\033[1m$1\033[0m"
 }
 
+# for action
+function bold_cyan(){
+	echo -e "\033[36m\033[1m$1\033[0m"
+}
 
 
-echo "ROOT $ROOT"
-echo "RESULT $RESULT"
 
-bold_magenta "clean $RESULT"
-rm $RESULT/*
+bold_magenta "ROOT   : $ROOT"
+bold_magenta "RESULT : $RESULT"
 
-
-bold_magenta "run c++ app"
-$ROOT/build/main
-breakline
-
-
-breakline
-bold_magenta "run python"
-python3 $ROOT/interface/testCore.py
-breakline
-
+if [ -d "$RESULT" ]; then
+	breakline
+	bold_cyan "clean $RESULT"
+	rm $RESULT/*
+fi
 
 breakline
-bold_magenta "render from blender"
-blender $ROOT/scene/scene.blend --python $ROOT/scene/testBScene.py --background
-breakline
+bold_cyan "c++ only"
+$ROOT/build/main -o $RESULT/c++
 
 
 breakline
-echo "listing $RESULT"
+bold_cyan "python"
+python3 $ROOT/interface/testCore.py -o $RESULT/py
 
+
+
+breakline
+bold_cyan "blender script"
+blender $ROOT/scene/scene.blend --python $ROOT/scene/testBScene.py --background -o $RESULT/bl_script
+
+
+breakline
+bold_cyan "blender engine"
+blender -b $ROOT/scene/scene.blend -o $RESULT/bl_engine# -f 0
+
+
+breakline
+bold_cyan "listing $RESULT"
 ls $RESULT
